subsystem: add MaskROM to RocketSubsystem (#2009)
* subsystem: add MaskROM to RocketSubsystem

* subsystem: factor bus topology out of example system
